+++ corsheaders/middleware.py
@@ -6,11 +6,7 @@ try:
 except ImportError:
     from urllib.parse import urlparse
 
-try:
-    from django.apps import apps
-    get_model = apps.get_model
-except ImportError:
-    from django.db.models.loading import get_model
+from django.apps import apps
 
 from corsheaders import defaults as settings
 
@@ -109,7 +105,7 @@ class CorsMiddleware(object):
             url = urlparse(origin)
 
             if settings.CORS_MODEL is not None:
-                model = get_model(*settings.CORS_MODEL.split('.'))
+                model = apps.get_model(*settings.CORS_MODEL.split('.'))
                 if model.objects.filter(cors=url.netloc).exists():
                     response[ACCESS_CONTROL_ALLOW_ORIGIN] = origin