+++ HISTORY.rst
@@ -4,8 +4,9 @@ History
 Pending
 -------
 
-* Drop Python 2.6 support
+* Drop Python 2.6 support.
 * Drop Django 1.3-1.7 support, as they are no longer supported.
+* Confirmed Django 1.9 support (no changes outside of tests were necessary).
 
 1.1.0 (2014-12-15)
 ------------------